+++ katana/api/action.py
@@ -16,6 +16,7 @@ from decimal import Decimal
 
 import zmq
 
+from ..logging import RequestLogger
 from ..payload import CommandPayload
 from ..payload import ErrorPayload
 from ..payload import get_path
@@ -248,6 +249,11 @@ class Action(Api):
             get_path(param, 'name'): Payload(param)
             for param in params
             }
+
+        # Logging is only enabled when debug is True
+        if self.is_debug():
+            rid = transport.get('meta/id')
+            self._logger = RequestLogger(rid, 'katana.api')
 
         service = self.get_name()
         version = self.get_version()