+++ katana/api/schema/file.py
@@ -9,15 +9,12 @@ For the full copyright and license information, please view the LICENSE
 file that was distributed with this source code.
 
 """
-import logging
 import sys
 
 from ...payload import Payload
 
 __license__ = "MIT"
 __copyright__ = "Copyright (c) 2016-2017 KUSANAGI S.L. (http://kusanagi.io)"
-
-LOG = logging.getLogger(__name__)
 
 
 class FileSchema(object):